Adds the possibility to add a local vim config
If you want to overwrite .vimrc settings add a "vimrc_local" file in the vim-dotfiles directory. My .vimrc will look if there's a vimrc_local and will load it at the end of the .vimrc.

This makes it possible to overwrite certain settings without having to add those settings to the git repository if they're only valid on your local installation.

The "vimrc_local" file is ignored by .gitignore.